refactor: 프로필 폴더 구조 변경 및 분리 (#157)
