+++ apps/web/app/(main)/profile/page.tsx
@@ -2,10 +2,10 @@ import { redirect } from 'next/navigation';
 
 import { Navigation, UserProfileCard } from '@/widgets/profile'; // UserProfileType import 추가
 
-import { getProfile } from '@/features/profile/api/getProfile';
 import { LogoutButton } from '@/widgets/authentication/LogoutButton';
 
 import { getCurrentUser } from '@/app/session';
+import { getProfile } from '@/shared/api/server/profile/getProfile';
 import { NavigationItem } from '@/shared/types/profile';
 
 const pointItems: NavigationItem[] = [
