+++ apps/web/shared/api/client/profile/useUpdateProfile.ts
@@ -1,7 +1,6 @@
+import { postProfile } from '@/shared/api/server/profile/postProfile';
 import { useMutation } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
-
-import { updateProfile } from '@/features/profile/api/updateProfile';
 
 interface UpdateProfileParams {
   id: string;
@@ -15,7 +14,7 @@ export const useUpdateProfile = () => {
   const router = useRouter();
 
   return useMutation({
-    mutationFn: (params: UpdateProfileParams) => updateProfile(params),
+    mutationFn: (params: UpdateProfileParams) => postProfile(params),
     onSuccess: () => {
       alert('프로필 정보가 성공적으로 업데이트되었습니다.');
       router.refresh();
