+++ apps/web/shared/test/profile/getProfile.test.ts
@@ -1,6 +1,6 @@
 import { supabase } from '@/shared/lib/supabase/supabase';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
-import { getProfile } from './getProfile';
+import { getProfile } from '../../api/server/profile/getProfile';
 
 // Mock Supabase client
 vi.mock('@/lib/supabase/supabase', () => ({
